refactor scripts to support  HDR PQ metrics & distributed processing
