+++ src/requirements.txt
@@ -1,0 +1,7 @@
+certifi==2021.5.30
+charset-normalizer==2.0.3
+idna==3.2
+requests==2.26.0
+urllib3==1.26.6
+numpy==1.19.1
+scipy==1.7.1
