+++ src/tasks.requirements.txt
@@ -1,0 +1,14 @@
+amqp==5.0.6
+billiard==3.6.4.0
+celery==5.1.2
+click==7.1.2
+click-didyoumean==0.3.0
+click-plugins==1.1.1
+click-repl==0.2.0
+kombu==5.1.0
+prompt-toolkit==3.0.21
+pytz==2021.3
+redis==3.5.3
+six==1.16.0
+vine==5.0.0
+wcwidth==0.2.5
